Fixed a bunch of errors, still not compiling

+++ src/stan/gm/arguments/arg_max_num_fp.hpp
@@ -23,7 +23,7 @@ namespace stan {
         _value = _default_value;
       };
       
-      bool is_valid(double value) { return value > 0; }
+      bool is_valid(int value) { return value > 0; }
 
     };
